API documentation

URNs

CitableBase.Urn — Type

Unique identifiers expressible in the syntax of the IETF's URN specification.

CitableBase.components — Function

components(uString)

Splits a string on colons (separator for top-level components of URNs).

Examples

julia> components("urn:cts:greekLit:tlg0012.tlg001.msA:1.1")

components(u)

Splits a URN's string representation into top-level components.

CitableBase.parts — Function

parts(componentString)

Splits a string on periods (seprator for parts within components of URNs).

Examples

julia> parts("tlg0012.tlg001.msA")

CitableBase.addversion — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

addversion(u, versioninfo)

CitableBase.dropversion — Function

Urn subtypes should implement dropversion(urn::U)::U.

dropversion(u)

Citable resources

CitableBase.Citable — Type

A citable unit of any kind is identified by a URN and has a human-readable label.

CitableBase.CitableTrait — Type

Abstraction of values for a citable trait.

CitableBase.NotCitable — Type

Value for the CitableTrait for everything not citable.

CitableBase.citabletrait — Function

Define default value of CitableTrait as NotCitable.

CitableBase.citable — Function

True if x is a citable object.

CitableBase.urn — Function

Delegate urn to specific functions based on type's citable trait value.

urn(x)

It is an error to invoke the urn function on material that is not citable.

urn(_, x)

CitableBase.label — Function

Delegate label to specific functions based on type's citable trait value.

It is an error to invoke the label function on material that is not citable.

label(_, x)

URN comparison

CitableBase.UrnComparisonTrait — Type

Abstraction of values for URN manipulation.

CitableBase.NotUrnComparable — Type

Value of the UrnComparisonTrait for evertything that can NOT be compared using URN logic.

CitableBase.urncomparisontrait — Function

Default value of UrnComparisonTrait is NotUrnComparable.

CitableBase.urncomparable — Function

True if type T implements the UrnComparisonTrait.

CitableBase.urnequals — Function

For subtypes of Urn, implement urnequals concretely.

urnequals(u1, u2)

URN-comparable objects must implement urnequals.

urnequals(x, y)

CitableBase.urncontains — Function

URN-comparable objects must implement urncontains.

CitableBase.urnsimilar — Function

URN-comparable objects must implement urnsimilar.

CEX serialization

CitableBase.CexTrait — Type

Trait for objects serializable in CEX format.

CitableBase.NotCexSerializable — Type

Value for the CexTrait for content not serializable to CEX format.

CitableBase.cextrait — Function

The default value of CexTrait is NotCexSerializable.

cextrait(_)

Subtypes of Citable are CexSerializable.

cextrait(_)

CitableBase.cexserializable — Function

True if type T is serializable to CEX format

CitableBase.cex — Function

Delegate cex to specific functions based on type's cextrait value.

It is an error to invoke the cex function on material that is not CEX serializable.

cex(, x; delimiter)

CitableBase.fromcex — Function

Dispatch fromcex on cextrait of T to instantiate an object of type T from CEX-formatted data cexsrc.

fromcex(cexsrc, T; delimiter, configuration, strict)

Implement fromcex using first string parameter for a file name.

fromcex(fname, T, freader; delimiter, configuration, strict)

Implement fromcex using first string parameter for a URL.

fromcex(url, T, ureader; delimiter, configuration, strict)

Implement fromcex using first string parameter for raw string data.

fromcex(cexsrc, T, freader; delimiter, configuration, strict)

It is an error to invoke the fromcex function on material that is not CEX serializable.

fromcex(, cex, T; delimiter, configuration, strict)

It is an error invoke fromcex function with an unimplemented trait value.

fromcex(traitvalue, cexsrc, T; delimiter, configuration, strict)

Citable collections

CitableBase.CitableCollectionTrait — Type

Abstraction of values for a citable library collection trait.

CitableBase.NotCitableCollection — Type

Value for the CitableCollectionTrait for evertything that is not a citable library collection.

CitableBase.citablecollectiontrait — Function

Define default value of CitableCollectionTrait as NotCitableCollection.

CitableBase.citablecollection — Function

True if x has the value CitableCollection for the CitableCollectionTrait.