API documentation

URNs

CitableBase.Urn — Type

Unique identifiers expressible in the syntax of the IETF's URN specification.

Concrete functions for Urns

CitableBase.components — Function

components(uString)

Splits a string on colons (separator for top-level components of URNs).

Examples

julia> components("urn:cts:greekLit:tlg0012.tlg001.msA:1.1")

components(u)

Splits a URN's string representation into top-level components.

CitableBase.parts — Function

parts(componentString)

Splits a string on periods (seprator for parts within components of URNs).

Examples

julia> parts("tlg0012.tlg001.msA")

CitableBase.urnmatches — Function

Urn subtypes should implement urnmatches(urn1::U, urn2::U)::Bool.

URN abstractions to implement for specific types

CitableBase.addversion — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

CitableBase.dropversion — Function

Urn subtypes should implement dropversion(urn::U)::U.

CitableBase.urncontains — Function

Urn subtypes should implement urncontains(urn1::U, urn2::U)::Bool

Citable resources

CitableBase.Citable — Type

A citable unit of any kind is identified by a URN and has a human-readable label.

Citable abstractions to implement for specific types

CitableBase.urn — Function

Citable content should implement urn(c::Citable)::Urn

CitableBase.label — Function

Citable content should implement label(c::Citable).

CitableBase.cex — Function

Citable content should implement cex(c::Citable).