Types

CitableBase.Urn — Type

Unique identifiers expressible in the syntax of the IETF's URN specification.

CitableBase.Citable — Type

A citable unit of any kind is identified by a URN and has a human-readable label.

Functions

CitableBase.validurn — Function

validurn(s)

True if s is valid for this type of URN.

This checks only the minimal generic syntax for a URN, and should be overriden to check more specific requirements of types extending URN.

Examples

julia> validurn("urn:isbn:1788998367")

CitableBase.components — Function

components(uString)

Splits a string on colons (separator for top-level components of URNs).

Examples

julia> components("urn:cts:greekLit:tlg0012.tlg001.msA:1.1")

components(u)

Splits a URN's string representation into top-level components.

CitableBase.parts — Function

parts(componentString)

Splits a string on periods (seprator for parts within components of URNs).

Examples

julia> parts("tlg0012.tlg001.msA")

Functions to implement for specific URN types

CitableBase.addversion — Function

Catch-all function for addversion method.

addversion(u, s)

CitableBase.dropversion — Function

Catch-all function for dropversion method.